Using the SkiaSharp package to compress and resize images as system.drawing is only usable on windows and the azure app service plan was switched to linux

+++ Application/Services/Interfaces/IImageCompressor.cs
@@ -8,6 +8,6 @@ namespace Application.Services.Interfaces
 {
     public interface IImageCompressor
     {
-        string CompressImage(string base64Image, long quality);
+        string CompressImage(string base64Image, int quality);
     }
 }
